+++ client/src/quic.rs
@@ -510,6 +510,8 @@ impl QuicClientStream {
         SubscribeStream::new(self.boxed())
     }
 
+    // clippy bug?
+    #[allow(clippy::missing_const_for_fn)]
     pub fn get_version(&self) -> &str {
         &self.version
     }