+++ filter/src/message.rs
@@ -1010,6 +1010,8 @@ impl MessageBlockMeta {
         }
     }
 
+    // clippy bug?
+    #[allow(clippy::missing_const_for_fn)]
     pub fn blockhash(&self) -> &str {
         match self {
             Self::Limited { block_meta, .. } => &block_meta.blockhash,