+++ richat/src/pubsub/notification.rs
@@ -4,6 +4,7 @@ use {
         pubsub::{solana::SubscribeMethod, SubscriptionId},
     },
     ::metrics::gauge,
+    agave_reserved_account_keys::ReservedAccountKeys,
     jsonrpsee_types::{SubscriptionPayload, SubscriptionResponse, TwoPointZero},
     richat_filter::message::{MessageBlock, MessageTransaction},
     richat_shared::five8::signature_encode,
@@ -12,7 +13,6 @@ use {
     solana_sdk::{
         clock::Slot,
         message::{v0::LoadedMessage, VersionedMessage},
-        reserved_account_keys::ReservedAccountKeys,
         transaction::TransactionVersion,
     },
     solana_transaction_status::{
@@ -260,6 +260,7 @@ impl RpcTransactionUpdate {
             loaded_addresses: OptionSerializer::Skip,
             return_data: OptionSerializer::Skip,
             compute_units_consumed: OptionSerializer::Skip,
+            cost_units: OptionSerializer::Skip,
         }
     }